pyproject.toml
[tool.poetry]
name = "otherapy"
version = "0.1.0"
description = "A voice-activated digital assistant for adults with Asperger's Syndrome."
authors = ["Juan Sugg Gilbert <juanpedrosugg@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
torch = "^1.10.0"
transformers = "^4.19.2"
insanely_fast_whisper = "^0.0.13"
TTS = "^0.5.0"
streamlit = "^1.9.0"
pydantic = "^1.9.0"
weaviate-client = "^3.2.0"
langchain = "^1.2.3"
asyncio = "^3.4.3"
weaviate = "^3.5.0"
sentence-transformers = "^2.3.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
